requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.10.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.38
python-dotenv==1.0.0
pyjwt==2.8.0
pydantic-settings==2.2.1